refactor: split live-state reading out of the subsystem declaration graph
graph.py now answers every question from the declarations alone and takes no AppState, which is the boundary deciding where a helper belongs. Availability probing and the capability and settings fingerprints move to liveness.py, leaving ordering and its validation together and both modules well inside the size budget.
